Form pages cleanup
