+++ next/pages/mestske-sluzby/dev/[slug].ts
@@ -1,9 +1,7 @@
 import { getFormDefinitionBySlugDev } from 'forms-shared/definitions/getFormDefinitionBySlug'
 import { VersionCompareContinueAction } from 'forms-shared/versioning/version-compare'
 
-import FormPageWrapper, {
-  FormPageWrapperWithContextProps,
-} from '../../../components/forms/FormPageWrapper'
+import FormPageWrapper, { FormPageProps } from '../../../components/forms/FormPage'
 import { makeSerializableFormDefinition } from '../../../components/forms/serializableFormDefinition'
 import { SsrAuthProviderHOC } from '../../../components/logic/SsrAuthContext'
 import { environment } from '../../../environment'
@@ -21,55 +19,54 @@ type Params = {
 /**
  * A route to preview forms in `forms-shared` folder. Backend functionality doesn't work. Works only in development.
  */
-export const getServerSideProps = amplifyGetServerSideProps<
-  FormPageWrapperWithContextProps & GlobalAppProps,
-  Params
->(async ({ context }) => {
-  if (!environment.featureToggles.developmentForms || !context.params) {
-    return { notFound: true }
-  }
+export const getServerSideProps = amplifyGetServerSideProps<FormPageProps & GlobalAppProps, Params>(
+  async ({ context }) => {
+    if (!environment.featureToggles.developmentForms || !context.params) {
+      return { notFound: true }
+    }
 
-  const { slug } = context.params
-  const formDefinition = getFormDefinitionBySlugDev(slug)
-  if (!formDefinition) {
-    return { notFound: true }
-  }
+    const { slug } = context.params
+    const formDefinition = getFormDefinitionBySlugDev(slug)
+    if (!formDefinition) {
+      return { notFound: true }
+    }
 
-  const { success: embeddedSuccess, isEmbedded } = handleEmbeddedFormRequest(
-    formDefinition,
-    context,
-  )
-  if (!embeddedSuccess) {
-    return { notFound: true }
-  }
+    const { success: embeddedSuccess, isEmbedded } = handleEmbeddedFormRequest(
+      formDefinition,
+      context,
+    )
+    if (!embeddedSuccess) {
+      return { notFound: true }
+    }
 
-  const initialFormDataJson = getDefaultFormDataForFormDefinition(formDefinition)
+    const initialFormDataJson = getDefaultFormDataForFormDefinition(formDefinition)
 
-  return {
-    props: {
-      formServerContext: {
-        formDefinition: makeSerializableFormDefinition(formDefinition),
-        formId: '',
-        initialFormDataJson,
-        initialServerFiles: [],
-        formSent: false,
-        initialSummaryJson: getInitialSummaryJson(
-          context.query,
-          formDefinition,
+    return {
+      props: {
+        formServerContext: {
+          formDefinition: makeSerializableFormDefinition(formDefinition),
+          formId: '',
           initialFormDataJson,
-        ),
-        formMigrationRequired: false,
-        isEmbedded,
-        isDevRoute: true,
-        strapiForm: { slug },
-        versionCompareContinueAction: VersionCompareContinueAction.None,
+          initialServerFiles: [],
+          formSent: false,
+          initialSummaryJson: getInitialSummaryJson(
+            context.query,
+            formDefinition,
+            initialFormDataJson,
+          ),
+          formMigrationRequired: false,
+          isEmbedded,
+          isDevRoute: true,
+          strapiForm: { slug },
+          versionCompareContinueAction: VersionCompareContinueAction.None,
+        },
+        appProps: {
+          externallyEmbedded: isEmbedded,
+        },
+        ...(await slovakServerSideTranslations()),
       },
-      appProps: {
-        externallyEmbedded: isEmbedded,
-      },
-      ...(await slovakServerSideTranslations()),
-    },
-  }
-})
+    }
+  },
+)
 
 export default SsrAuthProviderHOC(FormPageWrapper)
